requirements.txt
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.0.1
idna==3.4
Naked==0.1.32
pycryptodome==3.17
PyYAML==6.0
requests==2.28.2
shellescape==3.8.1
urllib3==1.26.14
wincertstore==0.2
